third_party/blink/web_tests/accessibility/aria-slider-increment-decrement.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<style>
  .rtl {
    direction: rtl;
  }
</style>
</head>
<body>
  <div id="slider"
       role="slider"
       aria-valuemin="0"
       aria-valuemax="10"
       aria-valuenow="5"></div>

<script>
  function checkEvent(event, expected_key) {
    if (expected_key == "ArrowUp") {
      assert_equals(event.code, "ArrowUp", "event.code on " + event.type);
      assert_equals(event.key, "ArrowUp", "event.key on " + event.type);
      assert_equals(event.keyCode, 38, "event.keyCode on " + event.type);
      assert_equals(event.which, 38, "event.which on " + event.type);
    } else if (expected_key == "ArrowDown") {
      assert_equals(event.code, "ArrowDown", "event.code on " + event.type);
      assert_equals(event.key, "ArrowDown", "event.key on " + event.type);
      assert_equals(event.keyCode, 40, "event.keyCode on " + event.type);
      assert_equals(event.which, 40, "event.which on " + event.type);
    } else if (expected_key == "ArrowLeft") {
      assert_equals(event.code, "ArrowLeft", "event.code on " + event.type);
      assert_equals(event.key, "ArrowLeft", "event.key on " + event.type);
      assert_equals(event.keyCode, 37, "event.keyCode on " + event.type);
      assert_equals(event.which, 37, "event.which on " + event.type);
    } else if (expected_key == "ArrowRight") {
      assert_equals(event.code, "ArrowRight", "event.code on " + event.type);
      assert_equals(event.key, "ArrowRight", "event.key on " + event.type);
      assert_equals(event.keyCode, 39, "event.keyCode on " + event.type);
      assert_equals(event.which, 39, "event.which on " + event.type);
    }

    assert_true(event.isTrusted, "event.isTrusted on " + event.type);
    assert_equals(event.charCode, 0, "event.charCode on " + event.type);
    assert_equals(event.target, slider, "event.target on " + event.type);
    assert_equals(event.srcElement, slider, "event.srcElement on " + event.type);
    assert_true(event.bubbles, "event.bubbles on " + event.type);
    assert_false(event.defaultPrevented, "event.defaultPrevented on " + event.type);
  }

  var slider = document.getElementById("slider");
  var axSlider = accessibilityController.accessibleElementById("slider");

  // Tests for horizontal, left-to-right slider

  promise_test(function(t) {
    return new Promise(resolve => {
      slider.addEventListener("keydown", resolve);
      axSlider.increment();
    }).then(event => {
      checkEvent(event, "ArrowRight");
    });
  }, "check that sending an increment event to an ARIA slider generates a right arrow keydown event");

  promise_test(function(t) {
    return new Promise(resolve => {
      slider.addEventListener("keyup", resolve);
      axSlider.increment();
    }).then(event => {
      checkEvent(event, "ArrowRight");
    });
  }, "check that sending an increment event to an ARIA slider generates a right arrow keyup event");

  promise_test(function(t) {
    return new Promise(resolve => {
      slider.addEventListener("keydown", resolve);
      axSlider.decrement();
    }).then(event => {
      checkEvent(event, "ArrowLeft");
    });
  }, "check that sending a decrement event to an ARIA slider generates a left arrow keydown event");

  promise_test(function(t) {
    return new Promise(resolve => {
      slider.addEventListener("keyup", resolve);
      axSlider.decrement();
    }).then(event => {
      checkEvent(event, "ArrowLeft");
    });
  }, "check that sending a decrement event to an ARIA slider generates a left arrow keyup event");

  // Tests for horizontal, right-to-left slider

  promise_test(function(t) {
    slider.classList.toggle("rtl");

    return new Promise(resolve => {
      slider.addEventListener("keydown", resolve);
      axSlider.increment();
    }).then(event => {
      checkEvent(event, "ArrowLeft");
    });
  }, "check that sending an increment event to a RTL ARIA slider generates a left arrow keydown event");

  promise_test(function(t) {
    return new Promise(resolve => {
      slider.addEventListener("keyup", resolve);
      axSlider.increment();
    }).then(event => {
      checkEvent(event, "ArrowLeft");
    });
  }, "check that sending an increment event to a RTL ARIA slider generates a left arrow keyup event");

  promise_test(function(t) {
    return new Promise(resolve => {
      slider.addEventListener("keydown", resolve);
      axSlider.decrement();
    }).then(event => {
      checkEvent(event, "ArrowRight");
    });
  }, "check that sending a decrement event to a RTL ARIA slider generates a right arrow keydown event");

  promise_test(function(t) {
    return new Promise(resolve => {
      slider.addEventListener("keyup", resolve);
      axSlider.decrement();
    }).then(event => {
      checkEvent(event, "ArrowRight");
    });
  }, "check that sending a decrement event to a RTL ARIA slider generates a right arrow keyup event");

  // Tests for vertical slider

  promise_test(function(t) {
    slider.setAttribute('aria-orientation', 'vertical');

    return new Promise(resolve => {
      slider.addEventListener("keydown", resolve);
      axSlider.increment();
    }).then(event => {
      checkEvent(event, "ArrowUp");
    });
  }, "check that sending an increment event to a vertical ARIA slider generates an up arrow keydown event");

  promise_test(function(t) {
    return new Promise(resolve => {
      slider.addEventListener("keyup", resolve);
      axSlider.increment();
    }).then(event => {
      checkEvent(event, "ArrowUp");
    });
  }, "check that sending an increment event to a vertical ARIA slider generates an up arrow keyup event");

  promise_test(function(t) {
    return new Promise(resolve => {
      slider.addEventListener("keydown", resolve);
      axSlider.decrement();
    }).then(event => {
      checkEvent(event, "ArrowDown");
    });
  }, "check that sending a decrement event to a vertical ARIA slider generates a down arrow keydown event");

  promise_test(function(t) {
    return new Promise(resolve => {
      slider.addEventListener("keyup", resolve);
      axSlider.decrement();
    }).then(event => {
      checkEvent(event, "ArrowDown");
    });
  }, "check that sending a decrement event to a vertical ARIA slider generates a down arrow keyup event");
</script>
</body>
</html>
